refactor(tests): refactor resource tests to be more granular
